split web api tests
